% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_codons_to_aa.R
\name{dna_codons_to_aa}
\alias{dna_codons_to_aa}
\title{This function reads codons (triplet-dna) as input,
and returns the corresponding amino acids for those codons}
\usage{
dna_codons_to_aa(codons)
}
\arguments{
\item{codons}{A string of codons.}
}
\value{
The amino acid for each codon.
}
\description{
This function reads codons (triplet-dna) as input,
and returns the corresponding amino acids for those codons
}
\examples{
dna_codons_to_aa(ATGTTAGATCGTTAAGCA)
dna_conds_to_aa(TATATGATTCTTGAATAT)
}
