% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_codons.R
\name{mk_codons}
\alias{mk_codons}
\title{Split a DNA sequence into codons}
\usage{
mk_codons(dna, s = 1)
}
\arguments{
\item{dna}{A string. This is the DNA sequence.}

\item{s}{An integer. Specifies the starting position of the first codon.}
}
\value{
A list of strings, representing a codon each.
}
\description{
Splits a string into substrings with 3 letters each. You can 
specify the starting position.
}
\examples{
mk_codons('atgttgggg', s=1)
mk_codons('cgatgttggggt', s=3)
}
