% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna}
\alias{random_dna}
\title{Given a size it will generate a random dna sequence}
\usage{
random_dna(l)
}
\arguments{
\item{l}{A whole number.}
}
\value{
Random sequence of DNA of size \code{l}.
}
\description{
Given a size it will generate a random dna sequence
}
\examples{
random_dna(100)
random_dna(5000)
}
