% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rna.R
\name{random_rna}
\alias{random_rna}
\title{Random RNA sequence generator.}
\usage{
random_rna(l)
}
\arguments{
\item{l}{An integer which determines the length of the generated RNA sequence.}
}
\value{
Returns an RNA sequence as a string.
}
\description{
\code{random_rna} generates a random RNA sequence of a certain length and
returns it as a string.
}
\examples{
random_rna(10)
random_rna(50000)
}
